move all tests to use the new config builder
