+++ packages/config/src/config-builder.test.ts
@@ -1,6 +1,7 @@
 import { describe, expect, it } from "vitest";
-import z from "zod/v4";
+import { z } from "zod/v4";
 import { ConfigBuilder, type ConfigBuilderOptions } from "./config-builder";
+import { defineConfigParser } from "./parser/define-config-parser";
 
 const baseConfigBuilderOptions: ConfigBuilderOptions = {
 	validate: (finalConfig, z) => {
@@ -16,6 +17,41 @@ const baseConfigBuilderOptions: ConfigBuilderOptions = {
 };
 
 describe("ConfigBuilder", () => {
+	it("should deep merge configuration", () => {
+		const schema = z.object({
+			foo: z.object({
+				bar: z.object({
+					zz: z.object({
+						test: z.object({
+							should_remain: z.string(),
+							another: z.boolean(),
+						}),
+					}),
+				}),
+			}),
+		});
+
+		expect(
+			getConfig({
+				validate: (config) => schema.parse(config),
+			})
+				.addSource(ConfigBuilder.fileSource("base-deep.json"))
+				.addSource(ConfigBuilder.fileSource("dev-deep.json"))
+				.build(),
+		).toEqual({
+			foo: {
+				bar: {
+					zz: {
+						test: {
+							should_remain: "123",
+							another: false,
+						},
+					},
+				},
+			},
+		});
+	});
+
 	it("should throw an error if no source was added", () => {
 		expect(() =>
 			new ConfigBuilder(baseConfigBuilderOptions).build(),
@@ -48,6 +84,15 @@ describe("ConfigBuilder", () => {
 		expect(() =>
 			getConfig()
 				// @ts-expect-error - testing invalid source
+				.addSource()
+				.build(),
+		).toThrowErrorMatchingInlineSnapshot(
+			"[Error: Invalid source. Please provide a valid one.]",
+		);
+
+		expect(() =>
+			getConfig()
+				// @ts-expect-error - testing invalid source
 				.addSource(new Date())
 				.build(),
 		).toThrowErrorMatchingInlineSnapshot(
@@ -62,6 +107,58 @@ describe("ConfigBuilder", () => {
 		).toThrowErrorMatchingInlineSnapshot(
 			"[Error: Invalid source. Please provide a valid one.]",
 		);
+	});
+
+	it("should re-throw parser error", () => {
+		const errorMessage = "This file is not valid";
+
+		expect(() =>
+			getConfig({
+				parser: defineConfigParser({
+					acceptedFileExtensions: ["json"],
+					parse() {
+						return {
+							ok: false,
+							error: new Error(errorMessage),
+						};
+					},
+				}),
+			})
+				.addSource(ConfigBuilder.fileSource("base.json"))
+				.build(),
+		).toThrowErrorMatchingInlineSnapshot("[Error: This file is not valid]");
+	});
+
+	it("should throw an error if file type is not accepted", () => {
+		expect(() =>
+			getConfig().addSource(ConfigBuilder.fileSource("base.txt")).build(),
+		).toThrowErrorMatchingInlineSnapshot(
+			`[Error: ".txt" file is not supported by this parser. Accepted files are: "json"]`,
+		);
+	});
+
+	it(`should provide a zod instance so I don't need to bring any external resource`, () => {
+		const config = getConfig({
+			validate: (finalConfig, z) => {
+				const schemaInsideValidate = z.object({
+					appURL: z.url(),
+					api: z.object({
+						port: z.coerce.number().int().positive(),
+					}),
+				});
+
+				return schemaInsideValidate.parse(finalConfig);
+			},
+		})
+			.addSource(ConfigBuilder.fileSource("base.json"))
+			.build();
+
+		expect(config).toEqual({
+			appURL: "https://my-site.com",
+			api: {
+				port: 3000,
+			},
+		});
 	});
 
 	describe("from json", () => {
@@ -168,16 +265,98 @@ describe("ConfigBuilder", () => {
 			delete process.env["TEST-api_-_port"];
 		});
 	});
+
+	describe("slots", () => {
+		function getConfig(options?: Partial<ConfigBuilderOptions>) {
+			const schema = z.object({
+				appURL: z.url(),
+				port: z.coerce.number().positive().int(),
+			});
+
+			return new ConfigBuilder({
+				validate: (finalConfig) => schema.parse(finalConfig),
+				configFolder: "./src/__fixtures__",
+				...options,
+			});
+		}
+
+		it("should replace the slotted value with the value from env var", () => {
+			process.env.PORT = "3000";
+
+			expect(
+				getConfig().addSource(ConfigBuilder.fileSource("slot.json")).build(),
+			).toEqual({
+				appURL: "http://localhost:3000",
+				port: 3000,
+			});
+
+			process.env.PORT = undefined;
+		});
+
+		it("should NOT replace the slotted value if env var is not defined", () => {
+			expect(
+				getConfig({
+					validate: (finalConfig) => finalConfig,
+				})
+					.addSource(ConfigBuilder.fileSource("slot.json"))
+					.build(),
+			).toEqual({
+				appURL: "http://localhost:$PORT",
+				port: "$PORT",
+			});
+		});
+
+		it("should handle multiple slots", () => {
+			process.env.PORT = "3000";
+
+			const schema = z.object({
+				appURL: z.url(),
+				port: z.coerce.number().positive().int(),
+				host: z.string(),
+			});
+
+			expect(
+				getConfig({
+					validate: (config) => schema.parse(config),
+				})
+					.addSource(ConfigBuilder.fileSource("slot-multiple.json"))
+					.build(),
+			).toEqual({
+				appURL: "http://localhost:3000",
+				port: 3000,
+				host: "$HOST",
+			});
+
+			process.env.PORT = undefined;
+		});
+
+		it("should accept custom slot prefix", () => {
+			process.env.PORT = "3000";
+
+			expect(
+				getConfig({
+					slotPrefix: "@",
+				})
+					.addSource(ConfigBuilder.fileSource("slot-prefix.json"))
+					.build(),
+			).toEqual({
+				appURL: "http://localhost:3000",
+				port: 3000,
+			});
+
+			process.env.PORT = undefined;
+		});
+	});
 });
 
-const schema = z.object({
-	appURL: z.url(),
-	api: z.object({
-		port: z.coerce.number().int().positive(),
-	}),
-});
+function getConfig(options?: Partial<ConfigBuilderOptions>) {
+	const schema = z.object({
+		appURL: z.url(),
+		api: z.object({
+			port: z.coerce.number().int().positive(),
+		}),
+	});
 
-function getConfig(options?: Omit<ConfigBuilderOptions, "validate">) {
 	return new ConfigBuilder({
 		validate: (finalConfig) => schema.parse(finalConfig),
 		configFolder: "./src/__fixtures__",
